feat:分数相关

+++ src/main/java/com/dd/ai_smart_course/dto/CoursesDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Data
@@ -22,7 +23,7 @@ public class CoursesDTO {
     private Timestamp updatedAt;
     private String statusStudent;
     private int StudentCount; // <-- 新增字段，用于存储学生数量
-    private int averageScore; // <-- 新增字段，用于存储平均分
+    private BigDecimal averageScore; // <-- 新增字段，用于存储平均分
 
 
 }
